Combine logging and short-circuiting build cache decorator
