+++ subprojects/core/src/main/java/org/gradle/caching/internal/BuildCacheServiceProvider.java
@@ -111,7 +111,6 @@ public class BuildCacheServiceProvider {
         RoleAwareBuildCacheService buildCacheService = new BuildCacheServiceWithRole(role, createRawBuildCacheService(buildCache));
         LOGGER.warn("Using {} as {} build cache, push is {}.", buildCacheService.getDescription(), role, buildCache.isPush() ? "enabled" : "disabled");
         buildCacheService = new BuildOperationFiringBuildCacheServiceDecorator(buildOperationExecutor, buildCacheService);
-        buildCacheService = new LoggingBuildCacheServiceDecorator(buildCacheService);
         buildCacheService = new ShortCircuitingErrorHandlerBuildCacheServiceDecorator(MAX_ERROR_COUNT_FOR_BUILD_CACHE, buildCacheService);
         return buildCacheService;
     }
