--- subprojects/core/src/test/groovy/org/gradle/caching/internal/LoggingBuildCacheServiceDecoratorTest.groovy
@@ -1,42 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.caching.internal
-
-import org.gradle.caching.BuildCacheService
-import spock.lang.Unroll
-
-class LoggingBuildCacheServiceDecoratorTest extends AbstractRoleAwareBuildCacheServiceDecoratorTest {
-    def decorator = new LoggingBuildCacheServiceDecorator(delegate)
-
-    BuildCacheService getDecorator() {
-        return decorator
-    }
-
-    @Unroll
-    def "does not suppress #exceptionType exceptions from store"() {
-        given:
-        delegate.store(key, writer) >> { throw exception }
-        when:
-        decorator.store(key, writer)
-        then:
-        thrown(exception.class)
-        where:
-        exception << exceptions
-        exceptionType = exception.class.simpleName
-    }
-
-}
